refactor: step-1 코드를 해당 폴더로 이동
