+++ __tests__/LottoReturnCalculator.test.js
@@ -1,8 +1,8 @@
-import LottoReturnCalculator from "../src/LottoReturnCalculator";
+import LottoReturnCalculator from '../src/step-1/LottoReturnCalculator.js';
 
-describe("LottoReturnCalculator 클래스 유닛테스트", () => {
-  describe("calculateReturnAmount", () => {
-    test("등수에 따른 수익 금액을 계산한다.", () => {
+describe('LottoReturnCalculator 클래스 유닛테스트', () => {
+  describe('calculateReturnAmount', () => {
+    test('등수에 따른 수익 금액을 계산한다.', () => {
       // given
       const rank = { 1: 1, 2: 0, 3: 0, 4: 0, 5: 0, 6: 0 };
 
@@ -14,8 +14,8 @@ describe("LottoReturnCalculator 클래스 유닛테스트", () => {
     });
   });
 
-  describe("calculateReturnRate", () => {
-    test("수익률을 계산한다.", () => {
+  describe('calculateReturnRate', () => {
+    test('수익률을 계산한다.', () => {
       // given
       const returnAmount = 2000000000;
       const purchaseAmount = 10000;
