+++ __tests__/utils.test.js
@@ -1,13 +1,10 @@
-import {
-  generateRandomNumber,
-  generateUniqueRandomNumbers,
-} from "../src/utils";
+import { generateRandomNumber, generateUniqueRandomNumbers } from '../src/step-1/utils.js';
 
-describe("utils 유닛테스트", () => {
-  describe("generateRandomNumber", () => {
-    test("1부터 주어진 값까지의 랜덤한 양의 정수를 반환한다.", () => {
+describe('utils 유닛테스트', () => {
+  describe('generateRandomNumber', () => {
+    test('1부터 주어진 값까지의 랜덤한 양의 정수를 반환한다.', () => {
       // given
-      jest.spyOn(Math, "random").mockReturnValueOnce(0.5);
+      jest.spyOn(Math, 'random').mockReturnValueOnce(0.5);
       const to = 45;
 
       // when
@@ -18,17 +15,17 @@ describe("utils 유닛테스트", () => {
     });
   });
 
-  describe("generateUniqueRandomNumbers", () => {
-    test("입력받은 길이의 1부터 주어진 값까지의 랜덤한 양의 정수 배열을 반환한다.", () => {
+  describe('generateUniqueRandomNumbers', () => {
+    test('입력받은 길이의 1부터 주어진 값까지의 랜덤한 양의 정수 배열을 반환한다.', () => {
       // given
       jest
-        .spyOn(Math, "random")
-        .mockReturnValueOnce(0.1)
-        .mockReturnValueOnce(0.3)
-        .mockReturnValueOnce(0.5)
-        .mockReturnValueOnce(0.7)
-        .mockReturnValueOnce(0.9)
-        .mockReturnValueOnce(0.2);
+      .spyOn(Math, 'random')
+      .mockReturnValueOnce(0.1)
+      .mockReturnValueOnce(0.3)
+      .mockReturnValueOnce(0.5)
+      .mockReturnValueOnce(0.7)
+      .mockReturnValueOnce(0.9)
+      .mockReturnValueOnce(0.2);
       const to = 45;
       const length = 6;
 
@@ -39,17 +36,17 @@ describe("utils 유닛테스트", () => {
       expect(result.length).toBe(6);
     });
 
-    test("입력받은 길이의 1부터 주어진 값까지의 랜덤한 중복되지 않은 숫자 배열을 반환한다.", () => {
+    test('입력받은 길이의 1부터 주어진 값까지의 랜덤한 중복되지 않은 숫자 배열을 반환한다.', () => {
       // given
       jest
-        .spyOn(Math, "random")
-        .mockReturnValueOnce(0.1)
-        .mockReturnValueOnce(0.3)
-        .mockReturnValueOnce(0.5)
-        .mockReturnValueOnce(0.7)
-        .mockReturnValueOnce(0.9)
-        .mockReturnValueOnce(0.9)
-        .mockReturnValueOnce(0.2);
+      .spyOn(Math, 'random')
+      .mockReturnValueOnce(0.1)
+      .mockReturnValueOnce(0.3)
+      .mockReturnValueOnce(0.5)
+      .mockReturnValueOnce(0.7)
+      .mockReturnValueOnce(0.9)
+      .mockReturnValueOnce(0.9)
+      .mockReturnValueOnce(0.2);
 
       const to = 45;
       const length = 6;
