+++ src/step-1/LottoReturnCalculator.js
